Taught extrude() how to work with different source & destination CAGs.
Previously extrude(cag, options) was limited to the "top" and "bottom"
of the extrusion (the CAGs defining the cross-section of the extrusion
in the plane normal to the offset vector) to be identical modulo
rotation. This change enables using different shapes for the top and
bottom, to e.g. effect a convex hull of a circle in the z=0 plane and
a square in the z=10 plane.

Previously discussed in
https://plus.google.com/112744897452366627400/posts/F9S8wXvyaVG

Some contrived examples: ![image](https://user-images.githubusercontent.com/114212/43186445-83e28fb0-8fa3-11e8-98c1-a63b0352eaa5.png)